Remove redundant plotting and add bin selection
Empirically, the resistantVirus population and the total virus population approach each other
near the end of the treatment regimen (That is, anything that wasn't resistant has already died,
so the only viruses which are left are resistant).  This means that recording both the resistant
virus population and the total virus population at the end of the trial yields no new information.
Accordingly, only the total virus population is recorded and plotted now.

Also, added the ability to specify the number of bins to be used for the histogram from the
function call.