Add function calls for answering questions
Part A asks several questions about the relationship of variables to the number of patients cured.
These functions are meant to answer those questions, and are meant (either) to be all uncommented
and run in sequence, or to be uncommented one at a time and run sequentially.

Note also that the graph title would need to be changed manually to reflect which trial is being run.