Add iteration and comments
The problem set asks about the effect of different delays between treatments with drugs,
so those delays have been included within a for loop that iteratively calls
simulationTwoDrugsDelayedTreatment.

Additionally, have dynamically titled the histograms to better track which simulation is plotted.

Finally, comment better through the function to see the purpose for the code.